refactor: StudyHistory, User Domain 관련 Service 분리
