+++ src/main/java/org/goormthon/beotkkotthon/rebook/service/studyhistory/ReadStudyHistoryListService.java
@@ -1,0 +1,40 @@
+package org.goormthon.beotkkotthon.rebook.service.studyhistory;
+
+import lombok.RequiredArgsConstructor;
+import org.goormthon.beotkkotthon.rebook.domain.StudyHistory;
+import org.goormthon.beotkkotthon.rebook.dto.request.StudyHistoryRequestDto;
+import org.goormthon.beotkkotthon.rebook.dto.response.studyhistory.StudyHistoryDetailDto;
+import org.goormthon.beotkkotthon.rebook.dto.response.studyhistory.StudyHistoryListDto;
+import org.goormthon.beotkkotthon.rebook.exception.CommonException;
+import org.goormthon.beotkkotthon.rebook.exception.ErrorCode;
+import org.goormthon.beotkkotthon.rebook.repository.StudyHistoryRepository;
+import org.goormthon.beotkkotthon.rebook.usecase.studyhistory.ReadStudyHistoryListUseCase;
+import org.goormthon.beotkkotthon.rebook.usecase.studyhistory.ReadStudyHistoryUseCase;
+import org.goormthon.beotkkotthon.rebook.usecase.studyhistory.UpdateStudyHistoryUseCase;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class ReadStudyHistoryListService implements ReadStudyHistoryListUseCase {
+    private final StudyHistoryRepository studyHistoryRepository;
+
+    @Override
+    public List<StudyHistoryListDto> execute(String category, Integer page, Integer size) {
+        Pageable pageable = PageRequest.of(page, size);
+        Page<StudyHistory> studyHistoryList = studyHistoryRepository.findAllByRecycleCategoryName(category, pageable);
+
+        return studyHistoryList.getContent().stream()
+                .map(studyHistory -> StudyHistoryListDto.builder()
+                        .id(studyHistory.getId())
+                        .imageUrl(studyHistory.getImageUrl())
+                        .isMarking(studyHistory.getIsMarking())
+                        .build())
+                .toList();
+    }
+}
