+++ src/main/java/org/goormthon/beotkkotthon/rebook/service/studyhistory/ReadStudyHistoryService.java
@@ -1,0 +1,37 @@
+package org.goormthon.beotkkotthon.rebook.service.studyhistory;
+
+import lombok.RequiredArgsConstructor;
+import org.goormthon.beotkkotthon.rebook.domain.StudyHistory;
+import org.goormthon.beotkkotthon.rebook.dto.request.StudyHistoryRequestDto;
+import org.goormthon.beotkkotthon.rebook.dto.response.studyhistory.StudyHistoryDetailDto;
+import org.goormthon.beotkkotthon.rebook.dto.response.studyhistory.StudyHistoryListDto;
+import org.goormthon.beotkkotthon.rebook.exception.CommonException;
+import org.goormthon.beotkkotthon.rebook.exception.ErrorCode;
+import org.goormthon.beotkkotthon.rebook.repository.StudyHistoryRepository;
+import org.goormthon.beotkkotthon.rebook.usecase.studyhistory.ReadStudyHistoryListUseCase;
+import org.goormthon.beotkkotthon.rebook.usecase.studyhistory.ReadStudyHistoryUseCase;
+import org.goormthon.beotkkotthon.rebook.usecase.studyhistory.UpdateStudyHistoryUseCase;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class ReadStudyHistoryService implements ReadStudyHistoryUseCase {
+    private final StudyHistoryRepository studyHistoryRepository;
+
+    @Override
+    public StudyHistoryDetailDto execute(Integer studyHistoryId) {
+        StudyHistory studyHistory = studyHistoryRepository.findById(studyHistoryId)
+                .orElseThrow(() -> new CommonException(ErrorCode.NOT_FOUND_RESOURCE));
+
+        return StudyHistoryDetailDto.builder()
+                .imageUrl(studyHistory.getImageUrl())
+                .content(studyHistory.getRecycleCategory().getDescription())
+                .build();
+    }
+}
