+++ src/main/java/org/goormthon/beotkkotthon/rebook/usecase/user/UpdateUserNotificationStatusUseCase.java
@@ -2,11 +2,10 @@ package org.goormthon.beotkkotthon.rebook.usecase.user;
 
 import org.goormthon.beotkkotthon.rebook.annotation.UseCase;
 import org.goormthon.beotkkotthon.rebook.dto.request.UserNotificationRequestDto;
-import org.goormthon.beotkkotthon.rebook.dto.response.user.UserDto;
 
 import java.util.UUID;
 
 @UseCase
-public interface UpdateUserNotificationUseCase {
+public interface UpdateUserNotificationStatusUseCase {
     Void execute(UUID userId, UserNotificationRequestDto userNotificationRequestDto);
 }
